requirements.txt
flask
flask_sqlalchemy
flask_bcrypt
flask_login
flask_wtf
sqlalchemy
requests
googlemaps
twilio
transformers
email-validator
langchain
langchain-community
python-dotenv
google-cloud-aiplatform
langchain-google-vertexai
python-dotenv
torch
sentence-transformers
faiss-cpu